Move storage of committees into a repository
